+++ packages/gamut-labs/src/AppHeaderMobile/AppHeaderMainMenuMobile/index.tsx
@@ -91,7 +91,7 @@ export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = (
         />
       ) : (
         <>
-          {renderSearch && renderSearch()}
+          {renderSearch?.()}
           {items.map((item) => mapItemToElement(item, action))}
         </>
       )}